modify files to organize
