+++ src/router.js
@@ -1,0 +1,7 @@
+import handlers from "./handlers.js";
+let router = {
+  "": handlers.index,
+  note: handlers.note,
+};
+
+export default router;
